:fire: REFACTOR: service/api/에 있는 UserService와 authService삭제 후 src/api/UserService내 getMonthlyData 이동
