--- src/services/api/authService.js
@@ -1,1 +1,0 @@
-// 인증 관련 API 호출 함수 모음 (예: 로그인, 로그아웃)
